+++ test/CanonTest.hs
@@ -18,5 +18,7 @@ tests =
           parseCommand ["model", "x.g4"] === CommandModel "x.g4"
           parseCommand ["check", "x.g4"] === CommandCheck "x.g4"
           parseCommand ["check"] === CommandUsage
+          parseCommand ["parse", "l.g4", "p.g4", "start", "f"] === CommandParse "l.g4" "p.g4" "start" "f"
+          parseCommand ["parse", "g.g4", "start", "f"] === CommandParseCombined "g.g4" "start" "f"
           parseCommand ["version"] === CommandVersion
     ]